ports/fupy: Remove hw includes (use litex upstream)
ports/fupy/hw were cached copies of the litex software includes;
however these change from time to time, which has caused the
CSR access includes to be out of date.  Instead of just caching
a new version, remove the "hw" directory, and build with an
environment that has a symlink to the litex includes.